+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -11,6 +11,7 @@ unreleased
 2.0.9 (2017-06-21)
 ++++++++++++++++++
 * `group deployment create`: Fixes issue where some parameter files were no longer recognized using @<file> syntax.
+* `resource\managedapp` commands: Support `--ids` argument.
 
 
 2.0.8 (2017-06-13)